# docs/admin/admin_swagger.yaml
basePath: /api
definitions:
...
  admin.PeerBatchDeleteForm:
    properties:
      row_ids:
        items:
          type: integer
        type: array
    required:
    - row_ids
    type: object
  admin.PeerForm:
    properties:
      alias:
        type: string
      cpu:
        type: string
      group_id:
        type: integer
      hostname:
        type: string
      id:
        type: string
      memory:
        type: string
      os:
        type: string
      row_id:
        type: integer
      username:
        type: string
      uuid:
        type: string
      version:
        type: string
    type: object
...
  model.Peer:
    properties:
      alias:
        type: string
      cpu:
        type: string
      created_at:
        type: string
      group_id:
        type: integer
      hostname:
        type: string
      id:
        type: string
      last_online_ip:
        type: string
      last_online_time:
        type: integer
      memory:
        type: string
      os:
        type: string
      row_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
      username:
        type: string
      uuid:
        type: string
      version:
        type: string
    type: object
  model.PeerList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Peer'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  model.StatusCode:
    enum:
    - 1
    - 2
    type: integer
    x-enum-comments:
      COMMON_STATUS_DISABLED: 通用状态 禁用
      COMMON_STATUS_ENABLE: 通用状态 启用
    x-enum-descriptions:
    - 通用状态 启用
    - 通用状态 禁用
    x-enum-varnames:
    - COMMON_STATUS_ENABLE
    - COMMON_STATUS_DISABLED
...
  model.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      group_id:
        type: integer
      id:
        type: integer
      is_admin:
        type: boolean
      nickname:
        type: string
      remark:
        type: string
      status:
        $ref: '#/definitions/model.StatusCode'
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: RustDesk API Server
  title: RustDesk API
  version: "1.0"
paths:
...
  /admin/peer/batchDelete:
    post:
      consumes:
      - application/json
      description: Suppression par lot d'appareils
      parameters:
      - description: ID de l'appareil
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.PeerBatchDeleteForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Suppression par lot d'appareils
      tags:
      - Appareil
  /admin/peer/create:
    post:
      consumes:
      - application/json
      description: Créer un appareil
      parameters:
      - description: Informations sur l'appareil
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.PeerForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Peer'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Créer un appareil
      tags:
      - Appareil
  /admin/peer/delete:
    post:
      consumes:
      - application/json
      description: Supprimer l'appareil
      parameters:
      - description: Informations sur l'appareil
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.PeerForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Supprimer l'appareil
      tags:
      - Appareil
  /admin/peer/detail/{id}:
    get:
      consumes:
      - application/json
      description: Détails de l'appareil
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Peer'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Détails de l'appareil
      tags:
      - Appareil
  /admin/peer/list:
    get:
      consumes:
      - application/json
      description: Liste des appareils
      parameters:
      - description: Numéro de page
        in: query
        name: page
        type: integer
      - description: Taille de la page
        in: query
        name: page_size
        type: integer
      - description: Temps
        in: query
        name: time_ago
        type: integer
      - description: ID
        in: query
        name: id
        type: string
      - description: Nom d'hôte
        in: query
        name: hostname
        type: string
      - description: uuids séparés par des virgules
        in: query
        name: uuids
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.PeerList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Liste des appareils
      tags:
      - Appareil
  /admin/peer/update:
    post:
      consumes:
      - application/json
      description: Modifier l'appareil
      parameters:
      - description: Informations sur l'appareil
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.PeerForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Peer'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Modifier l'appareil
      tags:
      - Appareil
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  token:
    in: header
    name: api-token
    type: apiKey
swagger: "2.0"